@@ -37,20 +37,19 @@
     </scm>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
 
         <junit.version>4.13.2</junit.version>
         <groovy.version>4.0.23</groovy.version>
         <spock.version>2.4-M1-groovy-4.0</spock.version>
-        <byte-buddy.version>1.15.2</byte-buddy.version>
+        <byte-buddy.version>1.15.11</byte-buddy.version>
         <guava.version>33.3.1-jre</guava.version>
         <hamcrest.version>2.2</hamcrest.version>
         <slf4j.version>2.0.16</slf4j.version>
         <cucumber.version>7.20.1</cucumber.version>
-        <cucumber.tag-expressions.version>6.1.1</cucumber.tag-expressions.version>
         <gson.version>2.11.0</gson.version>
-        <junit5.version>5.11.1</junit5.version>
+        <junit5.version>5.11.4</junit5.version>
         <mockito.version>3.3.3</mockito.version>
         <assertj.version>3.26.3</assertj.version>
         <opentest.version>1.3.0</opentest.version>
@@ -471,7 +475,7 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>3.13.0</version>
                     <configuration>
-                        <!-- put your configurations here -->
+                        <release>11</release>
                     </configuration>
                 </plugin>
                 <plugin>
