@@ -7,6 +7,7 @@ import io.restassured.filter.log.RequestLoggingFilter;
 import io.restassured.response.Response;
 import io.restassured.specification.FilterableRequestSpecification;
 import io.restassured.specification.FilterableResponseSpecification;
+import net.serenitybdd.rest.utils.RestRuntimeException;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -48,9 +49,9 @@ public class FieldsRecordingFilter implements Filter {
             this.recorded = this.recorded.replaceAll("\n$", "");
             return response;
         } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException("Incorrect implementation, should be used correct charset", e);
+            throw new RestRuntimeException("Incorrect implementation, should be used correct charset", e);
         } catch (IOException e) {
-            throw new RuntimeException("Some exception during recording fields", e);
+            throw new RestRuntimeException("Some exception during recording fields", e);
         }
     }
 
